requirements.txt
Mastodon.py
psutil